+++ docs/Changelog.md
@@ -8,8 +8,10 @@
 
 ### `1.5.0`
 
+ * Convenience send methods; text(), binary(), ping(), pong() (@sirn-se)
  * Opcode filter for receive() method (@sirn-se)
  * Fix for unordered framgemented messages (@sirn-se)
+ * Various code re-write (@sirn-se)
 
 ## `v1.4`